requirements.txt
Flask==2.2.5
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
pandas==2.0.3
scikit-learn==1.3.0
Werkzeug==2.2.3
h5py==3.10.0
imutils==0.5.4
scipy==1.11.4
tqdm==4.66.1
matplotlib==3.8.0